# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CodeVideo API",
    "description": "API for generating videos based on specified actions",
    "version": "1.0.0"
  },
  "paths": {
    "/create-video-v3": {
      "post": {
        "summary": "Create a video immediately (for testing purposes)",
        "description": "Create a video immediately with the specified actions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestBodyV3"
              }
            }
          }
        }
      }
    },
    "/create-video-immediately": {
      "post": {
        "summary": "Create a video immediately (for testing purposes)",
        "description": "Create a video immediately with the specified actions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Video created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "example": "https://coffee-app.sfo2.cdn.digitaloceanspaces.com/codevideo/47a2ad38-3364-4031-972a-2d7705de2d4a.mp4"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RequestBodyV3": {
        "type": "object",
        "properties": {
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IAction"
            }
          }
        }
      },
      "RequestBody": {
        "type": "object",
        "properties": {
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IAction"
            }
          },
          "language": {
            "$ref": "#/components/schemas/language"
          },
          "textToSpeechOption": {
            "$ref": "#/components/schemas/textToSpeechOption"
          }
        }
      },
      "IAction": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the action to perform",
            "example": "author-speak-before"
          },
          "value": {
            "type": "string",
            "description": "The value of the action",
            "example": "I'm going to log \"Hello, world!\" to the console."
          }
        }
      },
      "language": {
        "type": "string",
        "example": "javascript"
      },
      "textToSpeechOption": {
        "type": "string",
        "example": "festival"
      }
    }
  }
}